:construction: added multifill as an option in the dataset
replaced chart subType need from the colors
