+++ server/charts/AxisChart.js
@@ -3,31 +3,7 @@ const moment = require("moment");
 const NewBarChart = require("./NewBarChart");
 const NewLineChart = require("./NewLineChart");
 const NewPieChart = require("./NewPieChart");
-
-function determineType(data) {
-  let dataType;
-  if (data !== null && typeof data === "object" && data instanceof Array) {
-    dataType = "array";
-  }
-  if (data !== null && typeof data === "object" && !(data instanceof Array)) {
-    dataType = "object";
-  }
-  if (typeof data !== "object" && !(data instanceof Array) && typeof data === "boolean") {
-    dataType = "boolean";
-  }
-  if (typeof data !== "object" && !(data instanceof Array) && typeof data === "number") {
-    dataType = "number";
-  }
-  if (typeof data !== "object" && !(data instanceof Array) && typeof data === "string") {
-    dataType = "string";
-  }
-  if (typeof data !== "object" && !(data instanceof Array) && moment(data).isValid()
-    && ((typeof data === "number" && data.toString().length > 9) || (typeof data !== "number"))) {
-    dataType = "date";
-  }
-
-  return dataType;
-}
+const determineType = require("../modules/determineType");
 
 class AxisChart {
   constructor(data) {
